frontend-agent/requirements.txt
streamlit
requests
pydantic-settings
google-auth
google-auth-httplib2
requests
streamlit